pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "researchinc"
version = "0.1.0"
description = "ResearchInc application with FastAPI and websockets support"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "gunicorn>=20.1.0",
    "websockets>=11.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "anthropic",
    "sqlalchemy",
    "psycopg[binary]",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pytest_asyncio",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["researchinc"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
strict = true 